Changed how modeline is configured
See, I've now got this on a desktop and a laptop.

On the desktop, I don't want to see battery anything.

On the laptop, I have already run my battery dry once by not having a
batter meter.

So I did a 'site' lisp file like I have with my emacs; I can put
customizations there.

Note that stumptray must be the last thing to run.  I still have some
glitchiness with it associated with it trying to guess the space as
things (looking at you, Dropbox) put icons in the tray, but it fixes
itself once finished... thus far.