Modeline changes
First of all, I added CPU/memory display to the modeline.

Secondly, I put the left-hand and right-hand sides into variables
which are then passed to the modeline building function.  The idea is
that if I wanted to, I could add more functionality (like a battery
gauge) by concating to one of those variables in site.lisp.